put audit log into generic page
